requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coinbase-commerce==1.0.1
cryptography==3.4.7
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
mysqlclient==2.0.3
paypalrestsdk==1.13.1
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
stripe==2.58.0
urllib3==1.26.4
uuid==1.30
